requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
discord.py==2.4.0
frozenlist==1.4.1
idna==3.8
multidict==6.0.5
python-dotenv==1.0.1
yarl==1.9.4